requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
Flask>=3.0.3
requests>=2.32.3
requests-oauthlib>=2.0.0
Pillow>=10.3.0
qrcode[pil]>=7.4.2
waitress>=3.0.0
